requirements.txt
streamlit==1.27.2
Pillow==10.0.1
numpy==1.26.1
tensorflow==2.14.0
